refactor(testing/e2e): move _errors and client into e2e module; add deprecation shim and unit tests
- Revert workflow action ref from @fix/e2e-test-classes to @main
- Move typed error classes and AEWorkflowClient into e2e/_errors.py
  and e2e/client.py (canonical location); full_dag/_errors and
  full_dag/client now re-export from there for backward compat
- Add DeprecationWarning to full_dag/__init__ so consumers see the
  pending migration when they pull
- Update base.py and sql_app.py imports to use e2e._errors / e2e.client
- Add direct unit tests for FullDAGOutcome, build_seed_dag,
  build_agent_json, and SQLAppE2ETest
