+++ .github/workflows/e2e-full-reusable.yaml
@@ -177,7 +177,7 @@ jobs:
           fi
           echo "Using agent-name=$(grep '^name=' "$GITHUB_OUTPUT" | cut -d= -f2)"
 
-      - uses: atlanhq/application-sdk/.github/actions/sdr-e2e@fix/e2e-test-classes
+      - uses: atlanhq/application-sdk/.github/actions/sdr-e2e@main
         with:
           app-name: ${{ inputs.app-name }}
           app-image-name: ${{ inputs.app-image-name }}
