+++ application_sdk/testing/e2e/base.py
@@ -41,6 +41,13 @@ import orjson
 
 from application_sdk.contracts.types import ConnectionRef
 from application_sdk.observability.logger_adaptor import get_logger
+from application_sdk.testing.e2e._errors import (
+    ManifestDagMissingError,
+    ManifestFileNotFoundError,
+    MissingHarnessClassAttrError,
+    MissingHarnessEnvError,
+)
+from application_sdk.testing.e2e.client import AEWorkflowClient, DAGRunResult
 from application_sdk.testing.e2e.credential import CredentialBody
 from application_sdk.testing.e2e.payload import (
     AgentSpec,
@@ -49,13 +56,6 @@ from application_sdk.testing.e2e.payload import (
     build_ae_payload,
 )
 from application_sdk.testing.e2e.substitutions import MustacheSubstitutions
-from application_sdk.testing.full_dag._errors import (
-    ManifestDagMissingError,
-    ManifestFileNotFoundError,
-    MissingHarnessClassAttrError,
-    MissingHarnessEnvError,
-)
-from application_sdk.testing.full_dag.client import AEWorkflowClient, DAGRunResult
 
 logger = get_logger(__name__)
